Avoid expensive calls to edb:format() in "threads" request
Summary:
# Context
In the DAP specification, the "threads" request is used to retrieve the list of all threads/processes in the debuggee. Each "thread" has a user visible name, which for Erlang processes, includes the pid.

Because the debugger runs on a different node, if we were to convert a debuggee pid to a string, it would include in its name the "node" part (e.g. `<1234.567.0>` instead of `<0.567.0>`). To avoid this, we were using the `edb:format()` function, that is similar to `io_lib:format()` but the formatting runs on the debuggee.

# Problem
Because `edb:format()` runs on the debuggee node, it involves a remote call, message passing, etc. When debugging a node with tens of thousands of processes, this overhead is noticeable. To make things worse, because of copy-and-paste, we were calling `edb:format()` also for cases where the result would be the same when ran locally.

# This diff
We extend the `edb:process_info()` type so that one can also request a "string version of the process pid". Using this, we don't nee any additional remote calls.

Reviewed By: TheGeorge

Differential Revision: D72163055

fbshipit-source-id: 94d1a6dcc65599e8e212815c612e5aeb0699e830