Batch pid => threadId mappings
Summary:
# Context
The DAP specification wants to use numeric "threadIds" to identify "threads", which correspond to Erlang processes. This means that we need to keep a mapping between pids and these threadIds. For this, we have a dedicated gen_server that will given a pid will give us a new threadId, if it was never seen, or the previous one used.

# Problem
For the "threads" request, we need to get the list of processes, and map their pids to threadIds. So far, we were doing it one by one, which, when the number of processes is high, introduces a noticeable overhead

# This diff
We add a way to convert all the pids to threadIds in one operation, and use it to implement the "threads" request

Reviewed By: TheGeorge

Differential Revision: D72171773

fbshipit-source-id: e1fb01514155cfeffa863dceee54abedab1cfc78