Improve code organization in edb_server_break
Summary:
# Context
The `edb_server_break` module defines the data-structure used for breakpoint book-keeping, and the operations around this. By "book-keeping" here we mean mainly keeping track of which line-breakpoints were set in the VM, and why (e.g. requested by user; as way of implementing stepping on a pid, etc).. It also deals with the logic of deciding which breakpoints should pause execution, and which ones "false positives".

# Problem
This is a core, complex module. The fact that functions have historically added as they were needed, makes it currently a bit hard to follow.

# This diff
We organize the exported API and the implementation in sections, and lay out operations in a top-down way. This just moves code around, with no functional changes.

Reviewed By: RobinMorisset

Differential Revision: D72253170

fbshipit-source-id: 4abe8b51d1a7f694df3edec8ff845512e5a8dc01