Avoid stepping breakpoints on stack-frames with no exception handlers
Summary:
# Context
To implement the stepping functionality, we set breakpoints on all functions along the call-stack. For example, if we are stepping-over, we need:

  - Breakpoints on the top frame, to detect when the process is done stepping-over
  - Breakpoints on the frame of the caller, to detect when the function ends normally, or raises an exception that is caught by the caller
  - Breakpoints on every other function along the call stack, in case they are the ones that catch an exception

Because we want to avoid doing a control-flow analysis on the function AST, we have a very simple heuristic of setting breakpoints on every line.

# Problem
Breakpoints have a runtime cost, so we should try to minimize bogus breakpoints. In particular, we should avoid setting breakpoints on functions on a frame if we know that they won't catch exceptions; which we can do be inspecting the stack frames contents

# This diff
When setting step breakpoints, for each stack-frame along the call-stack, we classify stack-frames as "exception handling only" and otherwise and, in the former we don't set breakpoints when we don't find exception handlers for the frame.

In doing this, we generalize the previous mechanism to decide whether it was ok not to set breakpoints, that we use to report errors when stepping on code not compiled with beam_debug_info, etc.

Reviewed By: robertoaloi

Differential Revision: D72253169

fbshipit-source-id: fb914e337fb82253ab69f2e25f9dc16c239dcb39