Fix pid formatting in thread name
Summary:
# Context
In D72163055 we changed the formatting of the process "thread name", so that we rely on a string representation that comes as process info, instead of round-tripping again to the debuggee node for each process, to format it properly.

# Problem
Because we now get a "binary" representing the pid, we need to make sure to use `~s` instead of `~p` when further formatting. Because we failed to do it, we were showing the pid as a binary for processes with a registered name

# This diff
Fix the formatting

Reviewed By: robertoaloi

Differential Revision: D72305729

fbshipit-source-id: 926d8a4f263672618ae30bcf5276e6b3566ff8aa