Remove website
Summary:
We are not actively using the EDB website, so move all the current information to the README and remove it.

GitHub pages is already disabled, the `gh-pages` branch deleted, the homepage URL updated to point to https://hexdocs.pm/edb.

Eventual additional documentation can be added as "extra" ExDoc pages in [Hex.pm](https://hexdocs.pm/edb/readme.html)

Reviewed By: jcpetruzza

Differential Revision: D72387614

fbshipit-source-id: 30c5699171d28a26795a440f1e92d39119c8ba7b