Add function to find form by function to edb_server_code
Summary:
# Context
We are adding basic support for stepping-in. If we are standing on a call, we want to figure the call-target, as an MFA, using the debug_info AST of the module of the top frame. Then, we want to set breakpoints on that MFA.

# Problem
We already have a function in `edb_server_break` that can set breakpoints on an MFA, but it needs the abstract to check which lines need brekapoints.

# This diff
Add a utility to find a function by name, to complement the one to find a function by line.

Reviewed By: alanz

Differential Revision: D72318079

fbshipit-source-id: 61118cdb78b59013738ad81c204aea50058f20fe