Basic step-in implementation
Summary:
# Context
In the previous diffs we built the basic infrastructure needed to support some form of stepping-into a call.

# This diff
We add a `edb:step_in/1` function, with basic support for stepping in:

  - It handles static calls
  - It supports tail-calls and non-tail-calls

For the latter, since we are not doing tail-call analysis on the AST, we instead register two set of call-stack patterns we are expecting to see: one for the case of a tail call, where there is no frame for the current function, and one for the non-tail call where we do see the current frame.

Things still missing:

  - Comprehensions
  - Closures
  - Dynamic calls like `M:F(42)`

Reviewed By: alanz

Differential Revision: D72406299

fbshipit-source-id: e6438656f206ee229cac18c43ba5794decc657df