Implement stepIn request handler
Summary:
# Context
We have just extended edb to have a very simple support for stepping-in, but the handler for the stepIn request of the DAP server is not yet aware of it.

# This diff
We replace the dummy implementation of the stepIn handler by an actual call to `edb:step_in()`.

# Caveats
Notice that the implementation of `edb:step_in/1` is still very simple. Effectively, it will only work on calls like `foo(X)` or `foo:bar(X)`, but at the moment it will not work on more complex expressions like `Y = foo(X)`, etc.

Reviewed By: TheGeorge

Differential Revision: D72454010

fbshipit-source-id: 39d8fbe76bb70a8df90e88360d535cf6c1530378