Parse input to stepping requests
Summary:
# Context
The DAP is a JSON based protocol. In `edb_dap_transport` we decode the json messages into Erlang maps, and dispatch them to the appropriate handler.

Each handler, defines the type of the arguments it expects, following the DAP specification, and we get help from eqwalizer to check we are using the input correctly.

However, it is still possible for a client to send invalid input, which would make the handler crash, forcing us to debug things the handler, etc. So recently we introduced the option for each handler to define a "parser" for its input. In that way, if parsing fails, we return a proper error message, which makes things easier to diagnose.

# Problem
Most handlers don't have a proper parser at the moment, and instead accept whatever was given. We want to fix them as we work on them.

# This diff
Add parsers for all the stepping handlers. In doing this, we notice that the current semantic of type variables is not based on bounded polymorphism, the types for some of the parsing functions were useless, so need to fix those to be dynamic for now.

Reviewed By: TheGeorge

Differential Revision: D72450906

fbshipit-source-id: a18d4f619555e3d3404b837cf19aa78025553b92