Ignore frames with dynamic code when stepping
Summary:
# Context
When stepping, we try to set breakpoints on frames along the call-stack, to be able to stop when the execution leaves the current function, either because it ends, or because it raised an exception that was caught.

However, the call-stack may contain modules that were not built with beam_debug_info or other such issues, and we don't want to prevent stepping because of that.

# Problem
If one of the modules in the call-stack was dynamically compiled, and has no beam file, we will not be able to inspect the AST to figure out where to set breakpoints. We currently fail stepping, which is undesirable. A typical example of this is when the a mecked module is in the call-stack.

# This diff
If a frame in the call-stack is dynamically compiled and we are ok failing to set a breakpoint in this frame, just ignore it

Reviewed By: RobinMorisset

Differential Revision: D72581850

fbshipit-source-id: 8d443d9bcc26be8b77f90a3fc7bfd3d313aaf130