Temporarily ban certain modules from having breakpoints
Summary:
# Context
When stepping, we set breakpoint on all functions with exception handlers along the call-stack, and trigger them only when the pid stepping hit them (in the right call-stack context).

# Problem
If heavily used modules like gen_server, gen_factory, etc are in the call-stack (e.g., we are in the process of handling a call), then this breakpoint imposes a heavy overhead, as now every gen_server call needs to go through the debugger process, etc.

# This diff
As a temporary mitigation, we add a small list of modules that we will not try to set breakpoints on. Long term plan is to extend the VM with support for conditional breakpoints, so that the overhead becomes low.

Reviewed By: RobinMorisset

Differential Revision: D72639670

fbshipit-source-id: 42813d88c2c9769a5777363ce2e3f20bdc9ba915