Properly explain why a breakpoint could not be set
Summary:
# Context
When the client requests to set a breakpoint, the server calls edb:add_breakpoint() and in case of error, forwards the Erlang term to the client

# Problem
On VS Code side, when attempting to set a breakpoint fails, the "red dot" is replaced by a "grey dot" that one can hover to see the "error". At the moment, we can read things like {badkey, 2251}, which are not clear to the user

# This diff
We now provide a proper error message explaining what the problem is.

In the process, we update the documentation for `edb` regarding this, as the errors were forwarded from `erl_debugger` (where the explanation did exist)

Reviewed By: TheGeorge

Differential Revision: D72658319

fbshipit-source-id: 1f77aa94dc735d76b9b20198d5a44004058f008a