Support step-in for calls under match
Summary:
# Context
We started introducing support for "stepping-in". Atm, we depend on us figuring out the call target from the AST, so there are many cases to consider

# Problem
Right now we only detect cases where we are standing on a call. In particular, a basic case like this is not handled:

```
X = foo(42)
```

# This diff
We refactor slightly the logic to analyze the AST, so that we can search for sub-expressions. As a first case, we handle the `match_expr` case

Reviewed By: RobinMorisset

Differential Revision: D72721202

fbshipit-source-id: 7bcd75ba1017de1f72eb5dbe5b635d6936691a68