Improve step-in error when no target found
Summary:
# Context
To implement "step-in", given the current module and line, we analyze the AST, to find the expression on that line and figure out what is the call target. If we don't find a call, we return an error.

# Problem
We current return a `not_a_call` term (with the type of the current term), which makes sense only if we want to handle only the case where the expression is an application.

That would be the ideal solution if OTP was giving us an instrumentation point before every effectful expression, which may happen in the future but is not currently the case.

Instead, we wil need to do some kind of analysis to find calls under compound expression, which will tackle next. But in that case, `not_a_call` is misleading.

# This diff
We rename the error to `no_call_in_expr`, and still return the expression name, which makes it easier to troubleshoot.

In the process, introduce a type for cal-target errors, so we can rely on eqwalizer to find when we need to update the human-readable version of the message

Reviewed By: TheGeorge

Differential Revision: D72705948

fbshipit-source-id: c8f62de5b5c04427ffe741a9f99533202c728205