Load the module if necessary when setting a breakpoint
Summary:
# Context
When running a node in `interactive` mode, module will be lazily loaded the first time they are used. So, the question is what should one do when setting a breakpoint on a module that hasn't been loaded yet.

# Problem
At the moment, the core `edb` api will not load the module, which requires the DAP server to do it instead. This was complicating things, but also doesn't scale well: when stepping-in a call for a module that hasn't been loaded yet, we'd need to load the module too, but the DAP server doesn't know which module to load

# This diff
As a first step, we move the logic for ensuring the module of a an expliit breakpoint to the edb_server. This let us clean up the `set_breakpoints` handler of the DAP server

Reviewed By: Balajiganapathi

Differential Revision: D72968001

fbshipit-source-id: 21f5fa16dceab532781fe719e25cc87b0d265f3d