Handle preloaded modules gracefully
Summary:
# Context
The Erlang VM comes with a certain number of "preloaded" modules, that are available from boot. Examples are `erlang`, `init`, `persistent_term`, etc.

# Problem
For stepping, we need access to the AST of a module, so we need to find its beam file. For this we rely on `code:which()`, which for preloaded modules returns `preloaded`.

So far, we were failing at this point with a `preloaded` term, which is considered a runtime failure. This made things like trying to step in on a call to a `erlang` module fail with an exception, even if it is not a BIF involved

# This diff
We make the logic for handling `preloaded` modules more graceful. In particular, if `code:which()` tells us that a module is `preloaded`, we use `code:where_is_file()` to find the corresponding beam file.

Reviewed By: RobinMorisset

Differential Revision: D73050245

fbshipit-source-id: 845c26b95623b8874411fe60b1ad479e5a0443ce