Improve error message when trying to step in to an undefined function
Summary:
# Context
We are implementing support for stepping-in to calls. As part of it, we first find the target function, and then inspect the AST of its module, to find line location, etc.

# Problem
When we don't find the function in the AST, it means that the function does not exist. However, we were returning a more generic error in that case, that gets rendered as "Couldn't determine call target".

# This diff
We add a more specific error for the case when the function is not found and render it appropriately

Reviewed By: RobinMorisset

Differential Revision: D73047004

fbshipit-source-id: c1bbec0b124449432ee6b5c3ec5221b79ca98e9d