Load the module if necessary when stepping into a call
Summary:
# Context
When running a node in `interactive` mode, modules will be lazily loaded the first time they are used.

# Problem
It is possible that a process is currently stopped before a call to `foo:bar/3`, but module `foo` hasn't been loaded yet. In that case, trying to step-in will currently fail.

# This diff
When stepping in, preemptively load the module of the target. We introduce a new error for this case, and ensure that it is formatted appropriately in the DAP server

Reviewed By: RobinMorisset

Differential Revision: D73042792

fbshipit-source-id: fbfdb22c4792b8692a1bbdc18459153c990e7ddd