Support step-in in infix expressions
Summary:
# Context
We started introducing support for "stepping-in". Atm, we depend on us figuring out the call target from the AST, so there are many cases to consider

# Problem
We can't step-in on expressions like
```
X + foo(42)
```

# This diff
Handle infix expressions. These introduce the following problem:

* If we have `f(X) + g(Y)`, the order of execution is not determined.

So to avoid presenting surprises to the user, we decide not two step on anything in this case. Ideally the compiler will give us this info in the future and there will be no ambiguity.

In order to support this, we need to change the general logic a bit: we now try to find all possible call targets, and succeed only if we find exactly one.

Reviewed By: RobinMorisset

Differential Revision: D73525637

fbshipit-source-id: 802f01f7883a15907995124e0cbdf7edb931cd54