Expand/collapse variables
Summary:
# Context
For the `variables` DAP request, we currently return the value of the variable as the `value` field. The `variablesReference` property is always set to 0.
# Problem
For structured variables, such as lists, maps or the alike, the output could be very big and displaying the full value of the value as a binary can be impractical.
# Solution
For structured variables (maps, lists, tuples), return a positive `variablesReference`. This indicates to the client that the variable is structured and it will issue a new variables request to get its children. This allows us to leverage the VS Code UI to expand/collapse structured variables.

For the time being, only handle maps, tuples and lists are handled. Tuples and lists are 1-indexed, for consistency with the `element/1` and `lists:nth/2` Erlang functions.

Proplists and records are currently out of scope. For records, we'd need a way to fetch the corresponding definitions, while for proplists we'd need to check whether all elements are pairs. In both cases the effort does not seem worth.

Reviewed By: jcpetruzza

Differential Revision: D73186132

fbshipit-source-id: 641283c6deea2386ffec34fe486f178e0b5f04d2