Make step-in work when the target is ambiguous
Summary:
# Context
To impement step-in, we currently inspect the AST of the module, find the expression at the current line, and determine what call targets are there. When there is a single target, we know we need to put breakpoints there.

In cases where there are multiple targets on the same expression, as in `f(X) + g(Y)`, it is not defined whether `f/1` or `g/1`  will be ran first, so currently we fail to step-in here.

# Problem
For the user it is not the best experience.
A better UX would be to break on whichever of the two is executed first. Sure, the user won't be able to then step-in the other one (as the breakpoint on the return of the call will be on the next line, after the other function was executed), but it is preferrable to not allowing to step-in

# This diff
We lift the restriction that we should find a single call-target, and allow now multiple targets to be found. What we do then is add breakpoints in all of them, and whichever runs first, wins

Reviewed By: RobinMorisset

Differential Revision: D73589799

fbshipit-source-id: 5c262d7ae1e8f019b680fc74eac0bdcbd92adab9