Limit targets for step-in to calls in current line
Summary:
# Context
We started introducing support for "stepping-in". Atm, we depend on us figuring out the call target from the AST, so there are many cases to consider

# Problem
If we are on line X, and the top-level expression has subexpressions that are calls, but are on a different line, we allow to step-into those calls, which may be confusing.

Similarly, if we have something like:

```
f(X) +
    g(Y)
```

with the process paused on the top line, currently we don't let the user step-in given that it is ambiguous what the "next" call is, but the intention of the user at that point is very likely to be "step in f/1, as it is the function in the current line"

# This diff
We make it so that we only consider "candidate" targets that are on the current line. This means that in an example like this:

```
try
    foo:bar(X)
catch _:_ -> ...
end
```

if the current line corresponds to the `try`, it will not be possible to step-in, and one first needs to stop-over to the line with the foo:bar/2.

It also means that in:

```
f(X) +
  g(Y)
```

the user will now be able to step into f/1.

Reviewed By: RobinMorisset

Differential Revision: D73584603

fbshipit-source-id: d40b370fe4ea654a400998394b2fd69b8527bfef