Use erlang:breakpoints/3 for stepping
Summary:
# Context
When stepping-over or stepping-out, we need to add breakpoints on every line of functions accross the call-stack. The question is how to find out those line numbers.

At the moment, we rely on the AST that is included in the beam file when compiled with `+debug_info`.

# Problem
This has a couple of issues, but essentially it forces us to guess the beam file from which a module was loaded, and we risk setting breakpoints on lines that don't really correspond to the code that was actually loaded.

Moreover, it means that we have no way of stepping for modules that were dynamically loaded and don't have a file on disk backing them.

# This diff
The `erl_debugger` module now exposes a couple of BIFs to find out the breakable lines per module/function, so we now use that instead of the AST.

After this change, we still need to use the AST for stepping-in, to identify call-targets, which we aim to resolve at a later stage.

Reviewed By: RobinMorisset

Differential Revision: D73860980

fbshipit-source-id: 3653099949610617af735c27fe7ddde2def44475