Expand get_debug_info/1 to return calls info
Summary:
# Context
So far we were only projecting the var mapping information that we get from `code:get_debug_info/1`, since that was the only thing needed for building stack-frames.

# Problem
We now want to use the call-targets as well for stepping in.

# This diff
Ensure the call targets are also exposed and fix all call-sites

Reviewed By: robertoaloi

Differential Revision: D74261891

fbshipit-source-id: 1153fe107edbb7ccb9fbd59106f2c9caf1d0445a