Use call-targets provided by the VM for stepping-in
Summary:
# Context
So far, when we needed to step-into a function, we'd load the AST of the current module, and "parse it", to figure out the call-targets.

# Problem
Using the AST is problematic per se, as it forces us to guess the beam file for a module, and doesn't work on dynamically loaded code.

But for stepping-in, it is particularly bad, as "parsing" is redoing essentially a lot of the work that compiler has already done; needs to be maintained as the language changes and it is just hard to get right (e.g., at the moment we don't handle imported functions)

# This diff
We have extended the compiler so that it includes a list of call-targets per line in the new debug-info section, and we have adapted the code to use the new API.

Here, we finally make use of this, and rely only on the call targets provided by `code:get_debug_info/1` for stepping-in. In doing this, we can throw away a lot of code.

While the targets from `code:get_debug_info/1` provde enough information to handle dynamic calls (where the the target depends on variables), we are not using this information at the moment, and focus on static calls only. We will add that later.

On the other hand, with this change, we start handling list-comprehensions and closures, that before were not handled by the AST parser.

Reviewed By: robertoaloi

Differential Revision: D74261889

fbshipit-source-id: dce6f9dcd141d9a4ff86743ef606d02e75e72468