Move get_debug_info/1 function to edb_server_code
Summary:
# Context
Currently, we call `code:get_debug_info/1` only for the sake of building stack traces, via a wrapper function that is a private helper in `edb_server_stack_frames`.

# Problem
To implement step-in, we want to call `get_code_info/1` from the breakpoint handling module, but it doesn't make a lot of sense as a public function of `edb_server_stack_frames`.

# This diff
Move the function to `edb_server_code` and make it public

Reviewed By: robertoaloi

Differential Revision: D74261890

fbshipit-source-id: 7ab6b464a79635407ecd4d3d27550fc08a014503