Bump edb toolchain to fix bug in erl_debugger:breakpoints/2,3
Summary:
We bring a new version of the toolchain that fixes a bug where `erl_debugger:breakpoints/2,3` would segfault if the module we are querying has no line infornation at all.

This in particular is the case with the `erlang` module as compiled by OTP, and would lead to crashes when stepping-in on a line that had (hidden) calls to `erlang:error()` or `erlang:!()`.

Diff of changes [here](https://github.com/jcpetruzza/otp/compare/edb-2025-05-06..edb-2025-05-08)

Reviewed By: TD5

Differential Revision: D74402613

fbshipit-source-id: 987c44615927878e2610dfd1bb9d37c4325fc04d