Better handling of implicit function calls during step-in
Summary:
# Context
To implement step-into a function, we get the list of all possible call-targets at the current location, and then set breakpoints there. If we fail to set any breakpoint, we fail the full operation, as otherwise the user may get the false impression that some function was not really called.

To get the list of call targets, we now rely on the debug-info provided from the compiler, that is based on an analysis of the SSA code.

# Problem
The erlang compiler may introduce calls in the SSA that may not have been explicitly there. Typical examples are:

* Adding calls to `erlang:error/1` due to pattern matching failures
* Adding calls to `erlang:'!'/2` when `!` is used

Now, currently OTP modules are not compiled with debug-info, so trying to set a breakpoint on them will fail. This means that in something like:

```
Pid ! foo(42)
```

we will fail to step-into `foo/1` due to the additional call to `erlang:`!`/2`.

# This diff
We add a workaround. Essentially, while trying to set breakpoitns on all the targets, we ignore failures to set breakpoints on the `erlang` module, assuming that among the targets there is at least one that is not on the `erlang` module

Reviewed By: acw224

Differential Revision: D74407207

fbshipit-source-id: d687ad0ca93238338316f7877a83eea066895427