Remove references to the +debug_info option
Summary:
# Context
The `+debug_info` option tells the compiler to add a section with the AST of the module. This is intended to be used by the `int` debugger.

Until recently, edb was also relying on this AST for the sake of stepping, but this is no longer the case, and it uses on the new `+beam_debug_info` option

# Problem
We were still using `+debug_info` in tests and mentioning it in the docs

# This diff
Remove all references to `+debug_info``

Reviewed By: Balajiganapathi

Differential Revision: D74561382

fbshipit-source-id: 015144a00ec083d5137f3ea1045b6e4d6dda0a45