Remove dependency on syntax_tools
Summary:
# Context
It used to be that for stepping, edb would load the AST that comes with +debug_info. In order to analyze it, it would use the `syntax_tools` library

# Problem
We no longer use the AST and there are no usages left of this app. We don't want to assume that the debuggee has it installed.

# This diff
Remove the dependency

Reviewed By: Balajiganapathi

Differential Revision: D74561384

fbshipit-source-id: 99818472f5c1ec333cc108ee8b7ed645699fe689