Make it easier to raise exceptions in edb_server
Summary:
# Context
`edb_server` is the debugger process in the debuggee. In general, we don't want it to crash, if we can avoid it, as the debugging session is lost and everything resumes.

We currently have an ad-hoc mechanism to raise "invariant violations", that become exceptions on the caller side, without killing the server

# Problem
As we will start allowing more execution of client code in the debugger, we will like to have a more general mechanism of raising exception on the caller side (e.g., to apply the effect of an exception in caller provided code)

# This diff
Here we just generalize the infra we already have in place, with a generic `edb_server:raise()` mechanism, making `edb:invariant_violaition()` just use that.

`invariant_violation` exceptions are not handled in code, so the fact that we changed from `throw` to `error` on the caller side, doesn't make a meaningful difference

Reviewed By: RobinMorisset

Differential Revision: D74561387

fbshipit-source-id: d59f177b7efb213ea71b55d020f3854a849ca988