Refine types used for variables request
Summary:
# Context
In each DAP request handler, we define Erlang versions of the DAP types, and for those that can be sent as part of requests, we are introducing parsers.

Initially, we were defining fields that are "enums" as binaries, but we are now using atoms instead, since we get type errors if we produce an invalid response or match on the wrong value, etc.

# Problem
There are a couple of cases in the types for the "variables" request where we still use binaries.

One is for the request argument, even though we are already parsing them as atoms, so this was inconsistent and would have led to errors had we tried to use them.

The rest are responses, and in general are "open" types with recommended variables. Because we don;t plan to use our own values, we can just close them and limit them to the recommended ones (or eventually extend if we ever do(

Reviewed By: robertoaloi

Differential Revision: D74561380

fbshipit-source-id: 016032f53db2e290e9ecccb5be69a140fec88f1a