Refactor variables request handler for clarity
Summary:
# Context
The `variables` request handler returns all "variables" for a given scope, or, for some structured value.

 A scope, at the moment, can be:
  - "Locals", i,e, user variables
  - "Registers", like "Locals" but when we don't have variable-mapping information
  - 'Messages" the message queue of the process

We expect to add more scopes in the future

Structured values are maps, lists, etc, and a "variable" request can be used to obtain all their subparts

# Problem
This request has clearly many different responsibilities and at the moment, the logic is done all inline in a single large "handle" method, which makes it harder to identify the subparts.

# This diff
Let the "handle" function act singly as a dispatcher, and put each different responsibility in a dedicated handler. Hopefully, this will make it easier to follow

Reviewed By: RobinMorisset

Differential Revision: D74561388

fbshipit-source-id: 178d5dd89f59581822b7bcb93ebf6187b823b43e