Refine types used for scopes request
Summary:
# Context
In each DAP request handler, we define Erlang versions of the DAP types, and for those that can be sent as part of requests, we are introducing parsers.

Initially, we were defining fields that are "enums" as binaries, but we are now using atoms instead, since we get type errors if we produce an invalid response or match on the wrong value, etc.

# Problem
In the "scopes" request we were still using `binary()` instead of a proper enum in some cases, and had not input parser defined

# This diff
Use an enum, define a parser, and minor cleanups

Reviewed By: robertoaloi

Differential Revision: D74561385

fbshipit-source-id: d82a1378d8f80d5d56416102f151cc0ee6f9c260