Add a `edb:eval()` function
Summary:
# Context
We want to  be able to evaluate run erlang code in the context of the stack frame of a process. This will allow us to implement things such as a REPL, and watch expressions

# This diff
As a first step, we introduce a function that will evaluate a function reference on stack-frame vars. This handles:

  * Finding the stack frames
  * Timeouts
  * Exceptions
  * Getting killed
  * Loading the module of the function reference

The last point is tricky as the caller will typically be on a separate node. So the server uses an rpc call to retrieve the module from the caller.

Reviewed By: RobinMorisset

Differential Revision: D74561379

fbshipit-source-id: a53216139ba1129997d10416324f39b74605c382