Introduce the edb_dap_delegate module
Summary:
# Context
We have recently added a `edb:eval/1` function that let's the caller execute a callback that receives the stack-frame vars of a function. As a side effect, the module of the callback is transferred to the debuggee.

We want to use this mechanism in the implementation of the "scopes" and "variables" requests, so that the formatting of the variables happens in the debuggee and we avoid transferring large terms altogether

# Problem
* We want to minimize the number of modules of the DAP that need to be transferred to the debuggee
* We want to make it easier to audit that no dependencies other than kernel, stdlib and edb_core are used in the callback code (as we can't assume anything else to be present in the debuggee)

# This diff
Introduce a new edb_dap_delegate module, that will centralize all the calls to `edb:eval/1`, using only callbacks that will be defined in it. For this we introduce an opaque type, and later in the stack we'll add actual callbacks.

Reviewed By: robertoaloi

Differential Revision: D74561386

fbshipit-source-id: 9fd2d51bb7ca559fbb67db713bebfbcf6db290b3