Refactor scopes implementation to reduce roundtrips
Summary:
# Context
In the DAP spec, "scopes" are variable containers, such as "locals", "arguments", etc.

The "scopes" request returns a list of scopes available for the process/frame and, for each, a "variablesReference" to query the variables in the scope., and the client will then perform a "variables" request for each "variablesReference"

# Problem
At the moment there's a lot of round-tripping with the debuggee

  1. The "scopes" request is received: 2 roundtrips; one to get variable frames and decide if we show locals or registers; one to do rpc to compute the message queue len and decide if we show the "messages" scope
  2. For each variablesRef, one roundtrip to get the frame variables, and then, for each non-trivial var, a round-trip to format the value as a binary

There's an additional problem: currently we use a dedicated scope for the messages, and consider each message a "variable". This means that if we are paused on a process with a huge mailbox, we'll generate a gigantic number of variables

# This diff
We use the new `edb:eval()` infrastructure to resolve all the scopes in one roundtrip. Now we will get, all at once:
* All the scopes that need to be shown,
* With all their values,
* Including how to render them

Doing this, we can instead "cache" the precomputed scope variables and let the "variables" request just return that. This should make everything overall much more snappy.

In doing this, we turn the "Messages" scope into a "Process" scope, of which "Messages in queue" becomes a variable, with the count as value, but that can still be expanded to show the individual messages. We will later add the process dictionary and other metrics to this scope.

At this point, we are stil transferring the raw variable value as a return of computing the scopes, since we still need them to handle the "structured" variables (i.e., having a variablesRef to deal with elements of a list, map, etc). This is something we will solve later in the stack, as we make the "structured" variables also running on the debuggee.

Reviewed By: robertoaloi

Differential Revision: D74561383

fbshipit-source-id: e1984dd227adbf3c7fed2912cfc5b73dd6596e2d