Refactor frame_scope_or_structured() type
Summary:
# Context
The "variables" request is used in the DAP to get a list of variable values. This can be:
  - For a "scope" such as "locals", "registers", etc.
  - The substructure of a value, e.g. to list all the key values of a map.

Each "variables" request comes with a "variablesRef" id that identifies what the client is after. We use the id-mapper to map variablesRef to something that tells us what it is

# Problem
In the id-mapper, we associate variablesRef to something called "frame_scope_or_structured(), which is accurate but not very clear of what it means. And the contents of the type is  also accidental, as later in the stack we will no longer care about a "frame" altogether.

Moreover, we call it "var_reference" all around, which makes it look like it is a reference for a single variables, while it is for a set of variables.

# This diff
We rename the type to `variablesInfo`, and use the same representation we use elsewhere in the DAP server, of having a `type` tag in each map, to identify the summand.

We also ensure that "vars" is used consistently instead of "var"

Reviewed By: robertoaloi

Differential Revision: D74561381

fbshipit-source-id: b64b4c981b3f3cab44721f9c04df4083c7944928