Refactor "variables" implementation to avoid transferring large values
Summary:
# Context
For structured values (maps, lists, etc), the DAP server can show only a summary representation of the value, and offer the client to "expand" the view, to see the subelements in more detail.

# Problem
The way we have implemented this, we first transfer a full copy of the original value (which can be very large), and keep in in the DAP. Then, for each layer of the structure, we store a copy of the subelement. Because this happens accross processes, there is no subterm sharing, and can lead to O(n^2) space usage

# This diff
We levarage the recent addition of `edb:eval()` to avoid transferring values altogether. Instead, for each requested variable we now have:
  * A value representation
  * A function that let us navigate to that subterm on the debuggee node, if we ever need to access a subelement

This way, we only transfer as many representations are needed, but we only store paths, not full copies.

In doing this, we prepare to be able to query smaller "windows", so very large lists, etc can be processed as well efficiently. We will extend the DAP server to support this type of requests later in the stack

Reviewed By: robertoaloi

Differential Revision: D74593615

fbshipit-source-id: 762f714a2e68b758345583c4470b6465f25d1cb7