Remove edb:format()
Summary:
# Context
When displaying variable values, we want pids, port, etc to have values relative to the debuggee. For this we were using `edb:format()` that, essentially, runs `io_lib:format/2` over rpc

# Problem
This was super inefficient and involved a lot of rountripping. We have since changed the mechanism and we no longer need this function.

# This diff
Just get rid of it

Reviewed By: TheGeorge

Differential Revision: D74594020

fbshipit-source-id: 00f7a328a76de542ff18f316b4ebcbed31823bcf