Avoid unncessasry allocations computing windows of variables
Summary:
# Context
We have recently added the infrastructure so that we can support variables paging. Essentially, every want we want to compute the structure of a variable, we pass a "window" with "start" and "count" and first build a slice of the data based on that

# Problem
At the moment, the "count" has to be an integer, so it forces us to make a full copy of the data, even if we want to return it in full.

# This diff
We make the "count" be `non_neg_integer() | infinity`, and in  case `infinity` is given, reuse the data without copying it

Reviewed By: robertoaloi

Differential Revision: D74730190

fbshipit-source-id: 12c71e4a36dbe98ea94237d9d42940f06e129c13